:recycle: Moved name generating functions to a separate file
