+++ src/main.js
@@ -4,6 +4,7 @@ import { NumberInput } from './components/NumberInput/NumberInput'
 import { NameGenerator } from './components/NameGenerator/NameGenerator'
 import { TextArea } from './components/TextArea/TextArea'
 import { MessageBubble } from './components/MessageBubble/MessageBubble'
+import { generateUsername } from './lib/utils'
 // import { Icon } from './components/NameGenerator/Icon'
 
 const { div } = van.tags
@@ -14,7 +15,7 @@ const Hello = () =>
       { style: () => ` margin: 20px;width: 350px;` },
       NumberInput({ min: '1', name: 'example', alt: 'number input example' })
     ),
-    div({ style: () => ` margin: 20px;width: 350px;` }, NameGenerator),
+    div({ style: () => ` margin: 20px;width: 350px;` }, NameGenerator({ generateUsername })),
     div({ style: () => ` margin: 20px;width: 350px;` }, TextArea({ rows: '6', cols: '30' })),
     div({ style: () => ` margin: 20px;width: 350px;` }, MessageBubble({ text: 'Test', time: '2min', user: 'You' })),
     // div(Icon),
